moving course listing to its own class, renaming user interface to app
