+++ lmsgrabber/src/main/java/rpi/lmsgrabber/fxml/ui_mockup.fxml
@@ -1,9 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.web.*?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.PasswordField?>
@@ -14,7 +10,7 @@
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.web.WebView?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="rpi.lmsgrabber.LMSFxController">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/9" xmlns:fx="http://javafx.com/fxml/1" fx:controller="rpi.lmsgrabber.LMSFxController">
    <children>
       <TabPane prefHeight="400.0" prefWidth="600.0" tabClosingPolicy="UNAVAILABLE">
         <tabs>
@@ -27,7 +23,7 @@
                         <Label layoutX="44.0" layoutY="157.0" text="Password" />
                         <PasswordField fx:id="password" layoutX="42.0" layoutY="186.0" />
                         <WebView fx:id="web_preview" layoutX="227.0" layoutY="42.0" prefHeight="232.0" prefWidth="353.0" />
-                        <TextField fx:id="url" layoutX="42.0" layoutY="53.0" />
+                        <TextField fx:id="url" layoutX="42.0" layoutY="53.0" text="https://lms.rpi.edu/" />
                         <TextField fx:id="username" layoutX="42.0" layoutY="117.0" />
                         <Button fx:id="btn_grab" layoutX="44.0" layoutY="240.0" mnemonicParsing="false" onAction="#login_lms" prefHeight="25.0" prefWidth="145.0" text="Grab!" />
                      </children>
